Added RPC message timeout and cooldowntime for login function.
Reason for cooldowntime is that if in cluster setup whole server is down, it will put disk offline after first timeout from server. But it will not affect distributed lock RPC -requests, which does not see if disk is offline or online, those will still try to relogin on every request.

This will help with issues like: 5290, 5092 and 4887