Timeout for RPC calls, disk status messages and moving disks initial offline timestamp to be only setted when server is booting up.
Timeout:
If server does hard shutdown (like real life poweroff), there is no-one telling minio that the other end has gone down, and we should close exiting connections.
This timeout will help in those situations, and speedup the connection failing process.

Issue which I had was really similar to issues #5290 and #4887, this timeout fixes it.

Disk  messages:
It's nicer to debug when system tells if disk is marked offline or online. It's also nicer from sysadmins point of view.

Initial offline timestamp for disk:
It was illogical that the server had different timeperiods, when disks could not fail and it worked differently than in other timeperiods.